requirements.txt
pyyaml>=6.0
numpy>=1.26
pandas>=2.2
jsonschema==4.25.1
